Hotfix/issue 25 (#52)
* Issue #25 Added login with Stormpath

* Issue #25 Added logout

* Issue #25 Bumped Stormpath version

* Issue #25 Handling authorization

* Issue #25 Removed localhost

* Fixes #51 Removed unused code and renamed package

* MongoDB URI Removed by mistake

* Removed unused test
