--- src/main/java/hello/HelloRepository.java
@@ -1,9 +1,0 @@
-package hello;
-
-import org.springframework.data.mongodb.repository.MongoRepository;
-
-public interface HelloRepository extends MongoRepository<Hello, String> {
-
-    Hello findByMessage(String message);
-
-}
