+++ src/main/java/nephology/HelloController.java
@@ -1,4 +1,4 @@
-package hello;
+package nephology;
 
 import org.springframework.stereotype.Controller;
 import org.springframework.web.bind.annotation.RequestMapping;
@@ -11,6 +11,6 @@ public class HelloController {
 
     @RequestMapping(value = "/")
     public String index() {
-        return "index.html";
+        return "index";
     }
 }
