+++ src/main/java/nephology/Security.java
@@ -1,0 +1,18 @@
+package nephology;
+
+import org.springframework.context.annotation.Configuration;
+import org.springframework.security.config.annotation.web.builders.HttpSecurity;
+import org.springframework.security.config.annotation.web.configuration.WebSecurityConfigurerAdapter;
+
+import static com.stormpath.spring.config.StormpathWebSecurityConfigurer.stormpath;
+
+/**
+ * Created by PLGrubskiM on 2017-03-01.
+ */
+@Configuration
+public class Security extends WebSecurityConfigurerAdapter {
+    @Override
+    protected void configure(HttpSecurity http) throws Exception {
+        http.apply(stormpath());
+    }
+}
