--- src/main/java/org/fujitsu/nephology/App.java
@@ -1,14 +1,0 @@
-package org.fujitsu.nephology;
-
-/**
- * Hello world!
- *
- */
-// // TODO: 2017-01-19 Delete this class
-public class App 
-{
-    public static void main( String[] args )
-    {
-        System.out.println( "Hello World!" );
-    }
-}
