+++ src/main/resources/application.properties
@@ -2,4 +2,9 @@
 #   This file stores configuration properties for the application.
 #
 spring.data.rest.basePath=/api
+
 spring.data.mongodb.uri=mongodb://admin:nephology1234@ds135069.mlab.com:35069,ds135069.mlab.com:35069/nephology
+
+stormpath.application.href =
+stormpath.client.apiKey.id = 
+stormpath.client.apiKey.secret =
