+++ src/main/resources/templates/index.html
@@ -1,0 +1,29 @@
+<!DOCTYPE html>
+<html xmlns:th="http://www.thymeleaf.org">
+<head>
+    <title>React + Spring</title>
+</head>
+<body>
+<div id='root'></div>
+<div id='root2'></div>
+
+<script src="https://cdnjs.cloudflare.com/ajax/libs/react/15.3.2/react.min.js"></script>
+<script src="https://fb.me/react-dom-15.0.1.js"></script>
+<script src="https://cdnjs.cloudflare.com/ajax/libs/babel-core/5.8.24/browser.js"></script>
+<script src="https://cdnjs.cloudflare.com/ajax/libs/jquery/3.1.1/jquery.min.js"></script>
+<script src="https://cdnjs.cloudflare.com/ajax/libs/toastr.js/2.1.3/toastr.min.js"></script>
+
+<link rel="stylesheet" href="https://cdnjs.cloudflare.com/ajax/libs/toastr.js/2.1.3/toastr.min.css"/>
+<link rel="stylesheet" href="https://maxcdn.bootstrapcdn.com/bootstrap/3.3.7/css/bootstrap.min.css"/>
+
+<script type="text/babel" src="/public/app1.js"></script>
+<script type="text/babel" src="/public/app2.js"></script>
+
+<div class='container'>
+    <div id='logoutBtn'></div>
+    <form action="/logout" method="post">
+        <input class="btn btn-danger center-block" type="submit" value="Logout" />
+    </form>
+</div>
+</body>
+</html>
