+++ src/main/webapp/public/app2.js
@@ -1,0 +1,11 @@
+var HelloInput = React.createClass({
+  render: function() {
+    return (
+    <div>second component</div>
+    );
+  }
+});
+
+ReactDOM.render(
+  <HelloInput />, document.getElementById('root2')
+);
