--- src/test/java/org/fujitsu/nephology/AppTest.java
@@ -1,46 +1,0 @@
-package org.fujitsu.nephology;
-
-import hello.Application;
-import junit.framework.Test;
-import junit.framework.TestCase;
-import junit.framework.TestSuite;
-
-import static org.mockito.Mockito.mock;
-
-/**
- * Unit test for simple App.
- */
-//// TODO: 2017-01-29 Class to be removed - only for the initial configuration check
-public class AppTest
-    extends TestCase
-{
-    /**
-     * Create the test case
-     *
-     * @param testName name of the test case
-     */
-    public AppTest( String testName )
-    {
-        super( testName );
-    }
-
-    /**
-     * @return the suite of tests being tested
-     */
-    public static Test suite()
-    {
-        return new TestSuite( AppTest.class );
-    }
-
-    /**
-     * Rigourous Test :-)
-     */
-    public void testApp()
-    {
-        assertTrue( true );
-    }
-
-    public void testMockito() {
-        mock(Application.class);
-    }
-}
